requirements.txt
Adafruit-Blinka==8.1.0
adafruit-circuitpython-busdevice==5.1.10
adafruit-circuitpython-dht==3.7.4
adafruit-circuitpython-typing==1.7.1
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.26.0
Adafruit-PureIO==1.1.9
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
Pillow==9.2.0
pkg-resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
python-dotenv==0.20.0
pyusb==1.2.1
requests==2.28.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.3.0
urllib3==1.26.11
